Create delete_user_v1
This is the code for my delete user page. This is an admin page, where only Graeme and admin is able to access. Here, he is able to delete a user from the database. This is the first version of my delete user page for my music database. It only contains the necessary code to get the basic sketch / layout of the page. So this is just the build, where I've created the page and its basic structure.