Update and rename delete_user_v1 to delete_user_V2
In version 2 of my delete user page, I updated and changed the code to create more of an organized layout. Here, I created / defined sections so that it looked more like my design. Example I gave heights, width and paddings to a few sections so that they could show and stand out, etc. So now there was a proper layout showing and it looked a lot better.